// src/app/api/car-logbook/model/vehicle.ts
/**
 * Car Logbook Service REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { VehiclePower } from './vehiclePower';
import { Gasoline } from './gasoline';
import { LogbookEntry } from './logbookEntry';
import { LogbookFuelEntry } from './logbookFuelEntry';


export interface Vehicle { 
    /**
     * ID
     */
    id?: string;
    make?: string;
    model?: string;
    licensePlate?: string;
    yearOfConstruction?: string;
    power?: VehiclePower;
    gasoline?: Gasoline;
    photo?: string;
    primary?: boolean;
    lastLogbookEntry?: LogbookEntry;
    lastLogbookFuelEntry?: LogbookFuelEntry;
    currentMileage?: number;
    currentAge?: number;
}
export namespace Vehicle {
}